pyproject.toml
[tool.poetry]
name = "speller"
version = "0.1.2-dev"
description = "P300 speller based on LSL"
authors = ["Bernhard Stadlbauer <b.stadlbauer@gmx.net>"]
license = "GPL-3.0-only"
repository = "https://github.com/bstadlbauer/lsl-p300-speller"
readme = "README.md"


packages = [
    { include = "bstadlbauer", from = "src" },
]


[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.26.2"
pylsl = "^1.13.6"
Pillow = "^9.3"

[tool.poetry.dev-dependencies]
pre-commit = "^2.7.1"

[tool.poetry.scripts]
start-speller = 'bstadlbauer.p300speller.p300_speller:main'

[tool.black]
line-length = 120

[tool.isort]
profile = 'black'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
